Move the tenant secrets controller into its own package
It only runs in multi-tenant mode, so it is deleted wholesale on the OSS side. No core file names an Enterprise-only kind now.
